+++ src/rex.instrument/NEWS.rst
@@ -5,7 +5,7 @@ REX.INSTRUMENT Change Log
 .. contents:: Table of Contents
 
 
-0.13.0 (2015-01-xx)
+0.13.0 (2015-01-30)
 ===================
 
 - Added support for ``rex.setup`` v2.